+++ electron/commands/open-folder.ts
@@ -1,0 +1,9 @@
+import { shell } from "electron";
+import logit from "../utils/logit";
+
+const openFolder = async (event, payload) => {
+  logit("📂 Opening Folder: ", payload);
+  shell.openPath(payload);
+};
+
+export default openFolder;
