+++ electron/constants/models.ts
@@ -1,0 +1,9 @@
+const DEFAULT_MODELS = [
+  "realesrgan-x4plus",
+  "remacri",
+  "ultramix_balanced",
+  "ultrasharp",
+  "realesrgan-x4plus-anime",
+];
+
+export default DEFAULT_MODELS;
