+++ electron/get-device-specs.ts
