+++ electron/utils/slash.ts
@@ -1,0 +1,4 @@
+import { getPlatform } from "../get-device-specs";
+
+const slash: string = getPlatform() === "win" ? "\\" : "/";
+export default slash;
